Major views overhaul, new standard
