+++ app/core/templates/base/dashboard.html
@@ -39,7 +39,7 @@
                   {% endif %}
                   <button onclick="location.href='{% url 'discord_index' %}'" style="min-width: 75%;" class="btn btn-{% if request.user.discord %}success{%else%}danger{%endif%}">Discord</button>
                   <hr>
-                  <button onclick="location.href='https://forum.kryptedgaming.com'" style="min-width: 75%;" class="btn btn-{% if request.user.discourse %}success{%else%}danger{%endif%}">Forum</button>
+                  <button onclick="location.href='{{forum_url}}'" style="min-width: 75%;" class="btn btn-{% if request.user.discourse %}success{%else%}danger{%endif%}">Forum</button>
                   <hr>
                   <p>If a service has a red background, click on it to enable it.</p>
               </div>
